requirements.txt
beautifulsoup4==4.4.1
feedparser==5.2.1
Jinja2==2.8
MarkupSafe==0.23
python-dateutil==2.5.3
requests==2.10.0
six==1.10.0
